Move files out of nested src
